gnu: direnv: Remove go references from output.
In the long run, it would be good to move direnv across to the
go-build-system, and update the package, but these changes do succeed in
reducing the closure size a lot.

Thanks to Tanguy Le Carrour for raising this issue on guix-devel.

* gnu/packages/shellutils.scm (direnv)[arguments]: Include the
remove-go-references phase from the go-build-system.
[inputs,native-inputs]: Move all inputs to native-inputs, as these are no
longer referenced in the output.