lint: 'source' check no longer complains about unavailable mirrors.
Fixes a regression introduced in
50fc2384feb3bb2677d074f8f0deb5ae3c56b4d8.

Previously, 'guix lint -c source coreutils' would complain if one of the
mirrors was unavailable.  This is no longer the case.

* guix/lint.scm (check-source)[warnings-for-uris]: Use 'filter-map'.
Remove 'append-map' call.
Use 'append-map' here so that we can meaningfull compare the length or
URIS and that of WARNINGS.
Use '=' to compare lengths.