tests: Relax expectations for the 'home-page' checker.
Fixes a regression introduced in
50fc2384feb3bb2677d074f8f0deb5ae3c56b4d8.

* tests/lint.scm (warning-contains?): New procedure.
("home-page: host not found"): Use 'warning-contains?' instead of
testing for equality, as was the case before commit 50fc2384feb.
This handles the case where the 'getaddrinfo' error is not "Name or
service not known" but instead something like "System error" or
"Servname not supported for ai_socktype", as is the case in the build
environment.