gnu: mumimu: Remove reference to guile-email.
The use of guile-email in this package definition seems to cause errors when
working with other packages, errors like: error: guile-email: unbound variable

I'm not sure why the version of the Guile email package is appropriate here
either, but for now just to avoid the problems, inline the current version of
the package (and add a comment so this is more obvious). This issue was
probably introduced in: c7b2b539802eaa3f969e212c98eb671a1a75e9f3

* gnu/packages/mail.scm (mumimu)[version]: Inline the current value
of (package-version guile-email) as a string.