gnu: emacs: Patch references to /bin/sh.
When attempting to run ert-runner in the build side environment where /bin/sh
doesn't exist as part of the emacs-elpy test suite, Emacs was throwing the
error "Searching for program: No such file or directory, /bin/sh".  This is
because the default shell when SHELL is not set was "/bin/sh", as can be seen
in callproc.c:

  sh = getenv ("SHELL");
  Vshell_file_name = build_string (sh ? sh : "/bin/sh");

Thus, patch the occurrences of "/bin/sh" in the source files of Emacs to refer
to their correct store location.

* gnu/packages/emacs.scm (emacs)[phases]{patch-program-file-names}: Add phase.