machine: ssh: Fix guix deploy hang when using non-DCE UUIDs.
Fixes <https://issues.guix.gnu.org/issue/41237>.

The UUID type information was lost when passing to the lower gexp code strata,
which led to not being able to recreate the UUID in the generated script.
This occurred for non-DCE type UUIDs such as that of a FAT file system.

A following commit will prevent the find-partition-by-uuid procedure from
entering a loop when its UUID argument is invalid.

Reported-by: Brice Waegeneire <brice@waegenei.re>

* gnu/machine/ssh.scm (machine-check-file-system-availability): Replace the
STRING->UUID procedure by the UUID macro, and provide the UUID type as its
second argument.