gnu: mumble: Fix mumble-overlay.
The canonical mumble-overlay script searches for libmumble.so.1 in various
directories, none of which are relevant in Guix installations and
additionally does some work to match the architecture of the library to that
of the command that will be executed.
This work is unnecessary in Guix and only causes the script to fail, so we
instead just hardcode the path.

* gnu/packages/telephony.scm (mumble)[#:phases]: Add fix-mumble-overlay.

Signed-off-by: Ludovic Courtès <ludo@gnu.org>