gnu: notmuch: Update to 0.30-0.31rc1.
Note: This actually updates to 0.31~rc1, but here we instead use the version
number 0.30-0.31rc1 so that Guix will consider it older than version 0.31.
The rationale for updating to a release candidate is that earlier versions do
not work with Emacs 27.

* gnu/packages/mail.scm (notmuch): Update to 0.30-0.31rc1.
[source]: In the 'uri' field, avoid using the 'version' variable,
which doesn't match the upstream file name.  Add a 'file-name' field
to work around the fact that Guix does not allow tildes ("~").
[arguments]: Add 'disable-failing-tests' phase.